rename calib.filter into calib.filters
