+++ digicampipe/scripts/trigger_uniformity.py
@@ -24,7 +24,7 @@ import numpy as np
 from ctapipe.visualization import CameraDisplay
 from docopt import docopt
 
-from digicampipe.calib.filter import filter_event_types
+from digicampipe.calib.filters import filter_event_types
 from digicampipe.instrument.camera import DigiCam
 from digicampipe.instrument.geometry import compute_patch_matrix
 from digicampipe.io.event_stream import event_stream
